added a couple of tests
