+++ tests/test_combine.py
@@ -1,0 +1,16 @@
+"""Tests related to specific combine functionality"""
+
+from __future__ import annotations
+
+from fitscube.combine_fits import check_for_any_beam
+
+
+def test_check_for_any_beams_no_beams(file_list) -> None:
+    """See if we can confirm is all beams are in fits files"""
+    # file_list returns fits files without beam information
+    assert not check_for_any_beam(file_list=file_list)
+
+
+def test_check_for_any_beam_real_images(time_image_paths) -> None:
+    """See if beam is in any of these images"""
+    assert check_for_any_beam(file_list=time_image_paths)
